% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{galigor_conflicts}
\alias{galigor_conflicts}
\title{Conflicts between the galigor and other packages}
\usage{
galigor_conflicts()
}
\value{
No return value, called for side effects
}
\description{
This function lists all the conflicts between packages in the galigor
and other packages that you have loaded.
}
\details{
If dplyr is one of the select packages, then the following four conflicts
are deliberately ignored: \code{intersect}, \code{union}, \code{setequal},
and \code{setdiff} from dplyr. These functions make the base equivalents
generic, so shouldn't negatively affect any existing code.
}
\examples{
galigor_conflicts()
}
