% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{galigoR_deps}
\alias{galigoR_deps}
\title{List all galigoR dependencies}
\usage{
galigoR_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
galigoR packages.}
}
\description{
List all galigoR dependencies
}
