% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{galigoR_packages}
\alias{galigoR_packages}
\title{List all packages in the galigoR}
\usage{
galigoR_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include galigoR in the list?}
}
\description{
List all packages in the galigoR
}
\examples{
galigoR_packages()
}
