% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{galigoR_update}
\alias{galigoR_update}
\title{Update galigoR packages}
\usage{
galigoR_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
galigoR packages.}
}
\description{
This will check to see if all galigoR packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
galigoR_update()
}
}
