% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{galigor_deps}
\alias{galigor_deps}
\title{List all galigor dependencies}
\usage{
galigor_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
galigor packages.}
}
\value{
No return value, called for side effects
}
\description{
List all galigor dependencies
}
