% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{galigor_packages}
\alias{galigor_packages}
\title{List all packages in the galigor}
\usage{
galigor_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include galigor in the list?}
}
\value{
No return value, called for side effects
}
\description{
List all packages in the galigor
}
\examples{
galigor_packages()
}
