% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{galigor_update}
\alias{galigor_update}
\title{Update galigor packages}
\usage{
galigor_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
galigor packages.}
}
\value{
No return value, called for side effects
}
\description{
This will check to see if all galigor packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
galigor_update()
}
}
